Fixed multiple misplacement of new variables. Fixed few more casting warnings. Changed CANUSE flags can now be properly set on ITEMDEF and dynamic items (ITEMDEFs value is given to dynamic items on creation)

+++ 0.56.b/common/CRect.cpp
@@ -767,8 +767,8 @@ int CPointBase::StepLinePath( const CPointBase & ptSrc, int iSteps )
 	if ( ! iDist2D )
 		return 0;
 
-	m_x = ptSrc.m_x + IMULDIV( iSteps, dx, iDist2D );
-	m_y = ptSrc.m_y + IMULDIV( iSteps, dy, iDist2D );
+	m_x = static_cast<short>(ptSrc.m_x + IMULDIV( iSteps, dx, iDist2D ));
+	m_y = static_cast<short>(ptSrc.m_y + IMULDIV( iSteps, dy, iDist2D ));
 	return( iDist2D );
 }
